Baselines v6.2.0

+++ PostgreSQL.13/Tests/Linq/ConvertTests/Tests.Linq.ConvertTests.ToDefaultDecimal(PostgreSQL.13).sql
@@ -1,7 +1,7 @@
 ﻿-- PostgreSQL.13 PostgreSQL
 
 SELECT
-	(t."MoneyValue" * 1000)::decimal
+	t."MoneyValue" * 1000
 FROM
 	"LinqDataTypes" t
