+++ PostgreSQL.14/Tests/Linq/ConvertTests/Tests.Linq.ConvertTests.ToDefaultDecimal(PostgreSQL.14).sql
@@ -1,7 +1,7 @@
 ﻿-- PostgreSQL.14 PostgreSQL.13 PostgreSQL
 
 SELECT
-	(t."MoneyValue" * 1000)::decimal
+	t."MoneyValue" * 1000
 FROM
 	"LinqDataTypes" t
