+++ PostgreSQL.15/Tests/Linq/ConvertTests/Tests.Linq.ConvertTests.ToDefaultDecimal(PostgreSQL.15).sql
@@ -1,7 +1,7 @@
 ﻿-- PostgreSQL.15 PostgreSQL
 
 SELECT
-	(t."MoneyValue" * 1000)::decimal
+	t."MoneyValue" * 1000
 FROM
 	"LinqDataTypes" t
