+++ PostgreSQL.16/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateSecond(PostgreSQL.16).sql
@@ -1,7 +1,7 @@
 ﻿-- PostgreSQL.16 PostgreSQL.15 PostgreSQL
 
 SELECT
-	EXTRACT(EPOCH FROM ((t."DateTimeValue" + 100 * Interval '1 Minute')::timestamp - t."DateTimeValue"::timestamp))
+	EXTRACT(EPOCH FROM ((t."DateTimeValue" + 100 * Interval '1 Minute')::timestamp - t."DateTimeValue"::timestamp))::Float
 FROM
 	"LinqDataTypes" t
