+++ PostgreSQL.16/Tests/Linq/DateTimeOffsetTests/Tests.Linq.DateTimeOffsetTests.SubDateMillisecond(PostgreSQL.16).sql
@@ -1,7 +1,7 @@
 ﻿-- PostgreSQL.16 PostgreSQL.15 PostgreSQL
 
 SELECT
-	ROUND(EXTRACT(EPOCH FROM ((t."TransactionDate" + 1 * Interval '1 Second')::timestamp - t."TransactionDate"::timestamp)) * 1000)
+	(ROUND(EXTRACT(EPOCH FROM ((t."TransactionDate" + 1 * Interval '1 Second')::timestamp - t."TransactionDate"::timestamp)) * 1000))::Float
 FROM
 	"Transactions" t
