+++ PostgreSQL.17/Tests/Linq/ConvertTests/Tests.Linq.ConvertTests.ToDefaultDecimal(PostgreSQL.17).sql
@@ -1,7 +1,7 @@
 ﻿-- PostgreSQL.17 PostgreSQL.15 PostgreSQL
 
 SELECT
-	(t."MoneyValue" * 1000)::decimal
+	t."MoneyValue" * 1000
 FROM
 	"LinqDataTypes" t
